Add echo functionality
