--- src/main/java/Tyler.java
@@ -1,17 +1,0 @@
-public class Tyler {
-    public static void main(String[] args) {
-        String logo = "  _____      _           \n"
-                + " |_   _|   _| | ___ _ __ \n"
-                + "   | || | | | |/ _ \\ '__|\n"
-                + "   | || |_| | |  __/ |   \n"
-                + "   |_| \\__, |_|\\___|_|   \n"
-                + "       |___/             \n";
-        System.out.println("Hello from\n" + logo);
-        String separator = "____________________________________________________________\n";
-        String greeting = " Hello! I'm Tyler!\n"
-                + " What can I do for you?\n";
-        String farewell = " Bye-bye now!\n";
-        System.out.println(separator + greeting
-                + separator + farewell +separator);
-    }
-}
