+++ src/main/java/main/Tyler.java
@@ -1,0 +1,32 @@
+package main;
+
+import java.util.Scanner;
+
+public class Tyler {
+    public static void main(String[] args) {
+        String logo = "  _____      _           \n"
+                + " |_   _|   _| | ___ _ __ \n"
+                + "   | || | | | |/ _ \\ '__|\n"
+                + "   | || |_| | |  __/ |   \n"
+                + "   |_| \\__, |_|\\___|_|   \n"
+                + "       |___/             \n";
+        System.out.println("Hello from\n" + logo);
+
+        String separator = "\t"
+                + "____________________________________________________________\n";
+        String greeting = "\t" + " Hello! I'm Tyler!\n"
+                + "\t" + " What can I do for you?\n";
+        String farewell = "\t" + " Bye-bye now!\n";
+        System.out.print(separator + greeting
+                + separator);
+
+        Scanner sc = new Scanner(System.in);
+        String input = sc.nextLine();
+        while (!input.equals("bye")) {
+            System.out.print(separator);
+            System.out.print("\t " + input + "\n" + separator);
+            input = sc.nextLine();
+        }
+        System.out.println(separator + farewell + separator);
+    }
+}
